requirements.txt
gunicorn==20.1.0
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
Django==3.2.6
django-cors-headers==3.8.0
django-sslserver==0.22
djangorestframework==3.12.4
JSON-log-formatter==0.4.0
gevent==21.12.0

# opentelemetry packages
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-exporter-otlp==1.10.0
opentelemetry-sdk-extension-aws~=2.0
opentelemetry-propagator-aws-xray~=1.0


# Copied in from https://github.com/open-telemetry/opentelemetry-python-contrib/tree/main/instrumentation
# except for aws-lambda
opentelemetry-distro==0.29b0
opentelemetry-instrumentation-aiohttp-client==0.29b0
opentelemetry-instrumentation-aiopg==0.29b0
opentelemetry-instrumentation-asgi==0.29b0
opentelemetry-instrumentation-asyncpg==0.29b0
opentelemetry-instrumentation-boto==0.29b0
opentelemetry-instrumentation-botocore==0.29b0
opentelemetry-instrumentation-django==0.29b0
opentelemetry-instrumentation-httpx==0.29b0
opentelemetry-instrumentation-logging==0.29b0
opentelemetry-instrumentation-redis==0.29b0
opentelemetry-instrumentation-requests==0.29b0
opentelemetry-instrumentation-tornado==0.29b0
opentelemetry-instrumentation-urllib==0.29b0
opentelemetry-instrumentation-urllib3==0.29b0
opentelemetry-instrumentation-wsgi==0.29b0
